split tests (GetOrganizationTest)
